moved span dispatcherservlet.render from new plugin to existing spring-webvc
