--- apm-agent-plugins/apm-dispatcher-servlet-render/src/main/java/co/elastic/apm/agent/dispatcherservlet/package-info.java
@@ -1,23 +1,0 @@
-/*-
- * #%L
- * Elastic APM Java agent
- * %%
- * Copyright (C) 2018 Elastic and contributors
- * %%
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * #L%
- */
-@NonnullApi
-package co.elastic.apm.agent.dispatcherservlet;
-
-import co.elastic.apm.agent.annotation.NonnullApi;
