--- apm-agent-plugins/apm-dispatcher-servlet-render/src/main/resources/META-INF/services/co.elastic.apm.agent.bci.ElasticApmInstrumentation
@@ -1,1 +1,0 @@
-co.elastic.apm.agent.dispatcherservlet.DispatcherServletRenderInstrumentation
