+++ apm-agent-plugins/apm-spring-webmvc-plugin/src/main/resources/META-INF/services/co.elastic.apm.agent.bci.ElasticApmInstrumentation
@@ -1,1 +1,2 @@
 co.elastic.apm.agent.spring.webmvc.SpringTransactionNameInstrumentation
+co.elastic.apm.agent.spring.webmvc.DispatcherServletRenderInstrumentation
