+++ apm-agent-plugins/pom.xml
@@ -20,7 +20,6 @@
         <module>apm-httpclient-core</module>
         <module>apm-slf4j-plugin</module>
         <module>apm-es-restclient-plugin</module>
-        <module>apm-dispatcher-servlet-render</module>
     </modules>
 
     <properties>
